Compute GLB chunk size checks in 64-bit and validate material factor array lengths
GLBResourceReader::Init computed the GLB header/chunk size checks
(GLB_HEADER_BYTE_SIZE + jsonChunkLength, and the component-chunk size sum) using
32-bit arithmetic. For a large 32-bit JSON chunk length the addition can wrap,
producing a value smaller than the actual file length so the size check passes
incorrectly and ReadJson is then asked for a very large allocation. Perform these
additions in 64-bit (uint64_t) so the sums cannot wrap; the comparisons against
the 32-bit file length are unchanged in behavior for all valid files.

ParseMaterial read baseColorFactor[0..3] and emissiveFactor[0..2] by fixed index
without checking how many elements were parsed. When the caller disables schema
validation (SchemaFlags::DisableSchemaRoot), a shorter array would be read beyond
the elements that were provided. Validate the element count (4 and 3 respectively)
on the core parsing path, independent of schema flags, throwing InvalidGLTFException
otherwise.

Adds regression tests: a GLB with a 0xFFFFFFFF JSON chunk length is now rejected,
and short baseColorFactor/emissiveFactor arrays are rejected under
SchemaFlags::DisableSchemaRoot.

Copilot-Session: 039828de-4db2-48f8-a5de-c9f7d33baffa

+++ GLTFSDK/Source/GLBResourceReader.cpp
@@ -138,8 +138,9 @@ void GLBResourceReader::Init()
         throw InvalidGLTFException("Unsupported GLB Version: " + std::to_string(version));
     }
 
-    // Length has been validated as the actual length of the file, but make sure to include the header bytes in this check
-    if (length < (GLB_HEADER_BYTE_SIZE + jsonChunkLength))
+    // Perform the addition in 64-bit so it cannot wrap for large 32-bit chunk lengths: a wrapped sum could be
+    // smaller than the file length and pass this check incorrectly.
+    if (length < (static_cast<uint64_t>(GLB_HEADER_BYTE_SIZE) + jsonChunkLength))
     {
         throw InvalidGLTFException("File length " + std::to_string(length) + " less than content length " + std::to_string(jsonChunkLength) + 
             " plus header length " + std::to_string(GLB_HEADER_BYTE_SIZE));
@@ -148,7 +149,7 @@ void GLBResourceReader::Init()
     m_json = ReadJson(*m_buffer, jsonChunkLength);
 
     // If length is exactly equal to the json chunk length, plus the header, it means there is no binary buffer chunk
-    if (length == (GLB_HEADER_BYTE_SIZE + jsonChunkLength))
+    if (length == (static_cast<uint64_t>(GLB_HEADER_BYTE_SIZE) + jsonChunkLength))
     {
         return;
     }
@@ -161,8 +162,8 @@ void GLBResourceReader::Init()
         throw InvalidGLTFException("Binary chunk should appear second");
     }
 
-    // Verify that the sum of the sizes of the chunks (plus the headers) matches the size of the file
-    const uint32_t chunkSizeSum = GLB_HEADER_BYTE_SIZE + jsonChunkLength + sizeof(bufferChunkLength) + GLB_CHUNK_TYPE_SIZE +  bufferChunkLength;
+    // Computed in 64-bit so the sum cannot wrap in 32-bit and spuriously match the file length.
+    const uint64_t chunkSizeSum = static_cast<uint64_t>(GLB_HEADER_BYTE_SIZE) + jsonChunkLength + sizeof(bufferChunkLength) + GLB_CHUNK_TYPE_SIZE +  bufferChunkLength;
 
     if (chunkSizeSum != length)
     {
